now send mail when user comment

+++ config/setup.php
@@ -1,9 +1,9 @@
 <?php
-#block_all(); -> a mettre en dehors du serveur web
+if ( !isset($_GET['password']) || $_GET['password'] !== "abc2019")
+    exit();
 error_reporting(E_ALL);
 ini_set('display_errors', 1);
-require 'database.php';
-
+require_once($_SERVER['DOCUMENT_ROOT'].'/config/database.php');
 
 $statement =  $DB_connect->prepare("DROP DATABASE IF EXISTS db;");
 $statement->execute();
@@ -17,7 +17,8 @@ $statement =  $DB_connect->prepare("CREATE TABLE db.user
     email VARCHAR(255) CHARACTER SET utf8,
     password VARCHAR(255) CHARACTER SET utf8,
     token VARCHAR(255) CHARACTER SET utf8,
-    validated_account TINYINT
+    validated_account TINYINT,
+    want_mail  VARCHAR(255) DEFAULT 'yes'
 
   );");
 $statement->execute();
@@ -42,7 +43,7 @@ $statement =  $DB_connect->prepare("CREATE TABLE db.commentary
   (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     img_id INT NOT NULL,
-    body VARCHAR(255) CHARACTER SET utf8,
+    body LONGTEXT CHARACTER SET utf8,
     user_id INT NOT NULL
   );");
 $statement->execute();
